category 및 Bujeok 생성 기능 구현

+++ src/main/java/com/example/app/bujeok/entity/dto/BujeokCreateDto.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Builder
 @Getter
 public class BujeokCreateDto { // 프론트 -> controller -> service
-    private int category;
+    private Long category;
     private String content;
     private String cheerUp;
 }
